feat: add --no-install option
Adds parsing of command line options using yargs. The --no-install flag
can be specified to run the setup without leaving any footprint leading
back to this specific package. That is, this package will not be
installed as a dependency. A project configured this way will not
benefit from updates to the config, but will instead get a hard-coded
set of rules to start with. This is useful for external projects where I
want the benefit of an automatic configuration but don't want any ties
to my personal package.

Revamps test structure to allow for the different paths the script now
can take.

The ignore-file is read from disk instead of being duplicated in the
setup script. DRY!

Closes: #6